[css-grid] Avoid AvailableSize to figure out whether grid is indefinite
The new Baseline Alignment algorithm states that items with sizing
cyclic dependencies must be excluded from any baseline context they
participate in. One of these cyclic dependencies can happen with
intrinsic sized grid areas and relative items.

The grid spec states [1] that flex-sized tracks should be considered as
content-sized when the grid container has an indefinite size. We were
using the AvailableSize(direction) function to determine whether the grid
container is indefinite or not. However, this function doesn't provide
consistent results during the different phases of the grid layout logic.

The new Baseline Alignment logic is not integrated in the Grid Track
sizing algorithm. Hence, we need to ensure that an item that participates
in any baseline alignment context during the track sizing also does
during the alignment phase, at the end of the grid layout logic.

However, the AvailableSize(direction) could indicate that the grid has
an indefinite size during the track sizing, but it returns a definite
size during alignment. This issue leads to inconsistencies regarding the
participation of an item in baseline alignment, so it may raises the
assertion of having a context, like it happens in the bugs this CL tries
to fix.

[1] https://drafts.csswg.org/css-grid/#fr-unit

Bug: 867833, 874861
Change-Id: I668d399b920c9280a8e20b3e8362f562eded4770