Modify source path for wpt test of Accept-CH
This CL modifies source path that is used in wpt test of Accept-CH.
window.open() uses the base URL of the entry realm's window in order to
resolve a relative URL.
The entry realm should be main window on calling function
|acceptChLoaded| as a listener for load event of non-main window in this
test.
We want to open do_not_expect_client_hints_headers.html in resources/
so the relative URL in this test should start from "resources/".

Also, entry realm is not set correctly by current chromium
implementation, so fixed test is assumed to fail.

Change-Id: I973393e0a7ca74cac882eb58f5a58aa9a81556ff