[UserTiming] Correct measure where start is undefined
According to spec, if start is undefined for performance.measure(), 0 will be
used as the start time. However, the current implementation treats undefined as
"null".

This CL doesn't solve all the problems. For backward compatibility, L2 should
treat undefined passed to start/end as empty, null as a string 'null'. As a
temporary fix, we treat both null and undefined as undefined, which is safer
than as 'null'.

Bug: 876324

Change-Id: I7d74d6be9552d090ab7f28342be9a739017bf59b